NDArrayPool.cpp Change of pool allocation strategy for new arrays
previously the first array from the freelist was taken and reallocated
if the new array was smaller only the internal size was reduced w/o realloc
This can lead to a severe allocation overhead when working with
very large e.g. 4GB each and small arrays at the same time.
In the worst case every new small array is reusing a formely very large array.

The changed strategy searches the freelist for the array with the closest size.
If the closest size is more than 150% larger than the requested size a
reallocation is done. There are a lot of cases where a newly allocated
array size is equal to a formerly freed array size, where with the new
strategy the mem overhead is minimized.